backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# File Processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
chardet==5.2.0
python-magic==0.4.27

# Redis Cache
redis==5.0.1

# Payment Processing
stripe==7.8.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Logging
structlog==23.2.0

# HTTP Client (for external APIs)
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1